pyproject.toml
[project]
name = "nvd-claude-proxy"
version = "0.3.0"
description = "Run Claude Code (and any Anthropic SDK client) on NVIDIA NIM models via a local proxy."
requires-python = ">=3.11"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "nvd-claude-proxy contributors" }]
keywords = [
  "anthropic", "claude", "nvidia", "nim", "proxy",
  "llm", "openai-compatible", "claude-code", "nemotron",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Internet :: Proxy Servers",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Framework :: FastAPI",
]
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "httpx>=0.27",
  "pydantic>=2.9",
  "pydantic-settings>=2.6",
  "pyyaml>=6.0",
  "python-multipart>=0.0.12",
  "tiktoken>=0.8",
  "pillow>=11.0",
  "structlog>=24.4",
  "orjson>=3.10",
  "typer>=0.12",
  "rich>=13.0",
]

[project.urls]
"Homepage"    = "https://github.com/khiwn/nvd-claude-proxy"
"Bug Tracker" = "https://github.com/khiwn/nvd-claude-proxy/issues"
"Changelog"   = "https://github.com/khiwn/nvd-claude-proxy/releases"

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "pytest-httpx>=0.33",
  "ruff>=0.7",
  "mypy>=1.13",
  "respx>=0.22",
]
metrics = [
  "prometheus-client>=0.21",
]
pdf = [
  "pypdf>=4.0",
]
full = [
  "prometheus-client>=0.21",
  "pypdf>=4.0",
]

[project.scripts]
nvd-claude-proxy = "nvd_claude_proxy.main:run"
ncp = "nvd_claude_proxy.cli.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"nvd_claude_proxy" = ["py.typed"]
"nvd_claude_proxy.data" = ["*.yaml"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
